Upgrade all Maven plugins
Use strictly pluginManagement for settings versions in order to be able
to invoke single plugins and get the same version as specified in the
build, as well as configuration.

Executions, in addition to plugins part of the default (jar) lifecycle,
are configured in build/plugins.

Configure versions-maven-plugin to be able to check for newer
dependencies/plugin versions

Add maven-enforcer-plugin to set minimal Maven version required by
configured plugins.