Build on Travis on OpenJDK 8
oraclejdk8 only gives error:
Expected feature release number in range of 9 to 15, but got: 8
The command "~/bin/install-jdk.sh --target "/home/travis/oraclejdk8" --workspace "/home/travis/.cache/install-jdk" --feature "8" --license "BCL"" failed and exited with 3 during .

Supposedly it is possible to use oraclejdk8 if forcing the dist on
Travis to be "trusty", but it is EOL, so let's not start using something
which is already deprecated and will eventually go away. OpenJDK works
just fine.