Only do one build on Travis
Travis' default behaviour is to first do mvn install without tests to
"install" dependencies, and then execute a normal build with tests.
This is to align with the build lifecycle defined by Travis, but it is
not necessary for a Maven build, which should do everything it needs by
executing the idiomatic mvn clean install.

This should even further speed up the build.